requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
langchain
langgraph
mcp
structlog
python-dotenv